+++ android/app/src/main/java/com/zzang/chongdae/data/repository/CommentDetailRepositoryImpl.kt
@@ -1,6 +1,5 @@
 package com.zzang.chongdae.data.repository
 
-import com.zzang.chongdae.remote.mapper.toDomain
 import com.zzang.chongdae.data.source.comment.CommentRemoteDataSource
 import com.zzang.chongdae.domain.model.Comment
 import com.zzang.chongdae.domain.model.CommentOfferingInfo
@@ -8,6 +7,7 @@ import com.zzang.chongdae.domain.repository.CommentDetailRepository
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
 import com.zzang.chongdae.remote.dto.request.CommentRequest
+import com.zzang.chongdae.remote.mapper.toDomain
 
 class CommentDetailRepositoryImpl(
     private val commentRemoteDataSource: CommentRemoteDataSource,
