+++ android/app/src/main/java/com/zzang/chongdae/data/repository/CommentRoomsRepositoryImpl.kt
@@ -1,11 +1,11 @@
 package com.zzang.chongdae.data.repository
 
-import com.zzang.chongdae.data.remote.mapper.toDomain
 import com.zzang.chongdae.data.source.CommentRoomsDataSource
 import com.zzang.chongdae.domain.model.CommentRoom
 import com.zzang.chongdae.domain.repository.CommentRoomsRepository
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.mapper.toDomain
 
 class CommentRoomsRepositoryImpl(
     private val commentRoomsDataSource: CommentRoomsDataSource,
