+++ android/app/src/main/java/com/zzang/chongdae/data/repository/OfferingDetailRepositoryImpl.kt
@@ -1,12 +1,12 @@
 package com.zzang.chongdae.data.repository
 
-import com.zzang.chongdae.data.remote.dto.request.ParticipationRequest
-import com.zzang.chongdae.data.remote.mapper.toDomain
 import com.zzang.chongdae.data.source.OfferingDetailDataSource
 import com.zzang.chongdae.domain.model.OfferingDetail
 import com.zzang.chongdae.domain.repository.OfferingDetailRepository
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.dto.request.ParticipationRequest
+import com.zzang.chongdae.remote.mapper.toDomain
 
 class OfferingDetailRepositoryImpl(
     private val offeringDetailDataSource: OfferingDetailDataSource,
