+++ android/app/src/main/java/com/zzang/chongdae/data/source/CommentRoomsDataSource.kt
@@ -1,8 +1,8 @@
 package com.zzang.chongdae.data.source
 
-import com.zzang.chongdae.data.remote.dto.response.commentroom.CommentRoomsResponse
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.dto.response.commentroom.CommentRoomsResponse
 
 interface CommentRoomsDataSource {
     suspend fun fetchCommentRooms(): Result<CommentRoomsResponse, DataError.Network>
