+++ android/app/src/main/java/com/zzang/chongdae/data/source/OfferingDetailDataSource.kt
@@ -1,9 +1,9 @@
 package com.zzang.chongdae.data.source
 
-import com.zzang.chongdae.data.remote.dto.request.ParticipationRequest
-import com.zzang.chongdae.data.remote.dto.response.offering.OfferingDetailResponse
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.dto.request.ParticipationRequest
+import com.zzang.chongdae.remote.dto.response.offering.OfferingDetailResponse
 
 interface OfferingDetailDataSource {
     suspend fun fetchOfferingDetail(offeringId: Long): Result<OfferingDetailResponse, DataError.Network>
