+++ android/app/src/main/java/com/zzang/chongdae/data/source/ParticipantRemoteDataSource.kt
@@ -1,6 +1,5 @@
 package com.zzang.chongdae.data.source
 
-import com.zzang.chongdae.data.remote.dto.response.participants.ParticipantsResponse
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
 import com.zzang.chongdae.remote.dto.response.participants.ParticipantsResponse
