+++ android/app/src/main/java/com/zzang/chongdae/data/source/offering/OfferingRemoteDataSource.kt
@@ -1,13 +1,12 @@
 package com.zzang.chongdae.data.source.offering
 
-import com.zzang.chongdae.data.remote.dto.request.OfferingWriteRequest
-import com.zzang.chongdae.data.remote.dto.response.offering.FiltersResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.MeetingsResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.OfferingsResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.ProductUrlResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.RemoteOffering
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.dto.request.OfferingWriteRequest
+import com.zzang.chongdae.remote.dto.response.offering.FiltersResponse
+import com.zzang.chongdae.remote.dto.response.offering.MeetingsResponse
+import com.zzang.chongdae.remote.dto.response.offering.OfferingsResponse
+import com.zzang.chongdae.remote.dto.response.offering.ProductUrlResponse
 import com.zzang.chongdae.remote.dto.response.offering.RemoteOffering
 import okhttp3.MultipartBody
