+++ android/app/src/main/java/com/zzang/chongdae/local/source/CommentLocalDataSourceImpl.kt
@@ -1,8 +1,8 @@
 package com.zzang.chongdae.local.source
 
+import com.zzang.chongdae.data.source.comment.CommentLocalDataSource
 import com.zzang.chongdae.local.dao.CommentDao
 import com.zzang.chongdae.local.model.CommentEntity
-import com.zzang.chongdae.data.source.comment.CommentLocalDataSource
 
 class CommentLocalDataSourceImpl(private val commentDao: CommentDao) : CommentLocalDataSource {
     override suspend fun insertComments(comments: List<CommentEntity>) {
