+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/home/HomeFragment.kt
@@ -25,9 +25,9 @@ import com.google.firebase.analytics.FirebaseAnalytics
 import com.zzang.chongdae.ChongdaeApp
 import com.zzang.chongdae.ChongdaeApp.Companion.dataStore
 import com.zzang.chongdae.R
-import com.zzang.chongdae.local.source.UserPreferencesDataStore
 import com.zzang.chongdae.databinding.FragmentHomeBinding
 import com.zzang.chongdae.domain.model.FilterName
+import com.zzang.chongdae.local.source.UserPreferencesDataStore
 import com.zzang.chongdae.presentation.util.FirebaseAnalyticsManager
 import com.zzang.chongdae.presentation.view.MainActivity
 import com.zzang.chongdae.presentation.view.home.adapter.OfferingAdapter
