+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/login/LoginActivity.kt
@@ -13,8 +13,8 @@ import com.kakao.sdk.common.model.ClientErrorCause
 import com.kakao.sdk.user.UserApiClient
 import com.zzang.chongdae.ChongdaeApp
 import com.zzang.chongdae.ChongdaeApp.Companion.dataStore
-import com.zzang.chongdae.local.source.UserPreferencesDataStore
 import com.zzang.chongdae.databinding.ActivityLoginBinding
+import com.zzang.chongdae.local.source.UserPreferencesDataStore
 import com.zzang.chongdae.presentation.util.FirebaseAnalyticsManager
 import com.zzang.chongdae.presentation.view.MainActivity
