+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/mypage/MyPageFragment.kt
@@ -10,8 +10,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.google.firebase.analytics.FirebaseAnalytics
 import com.zzang.chongdae.ChongdaeApp.Companion.dataStore
-import com.zzang.chongdae.local.source.UserPreferencesDataStore
 import com.zzang.chongdae.databinding.FragmentMyPageBinding
+import com.zzang.chongdae.local.source.UserPreferencesDataStore
 import com.zzang.chongdae.presentation.util.FirebaseAnalyticsManager
 import com.zzang.chongdae.presentation.view.login.LoginActivity
