+++ android/app/src/main/java/com/zzang/chongdae/remote/api/CommentApiService.kt
@@ -1,10 +1,10 @@
 package com.zzang.chongdae.remote.api
 
-import com.zzang.chongdae.data.remote.dto.request.CommentRequest
-import com.zzang.chongdae.data.remote.dto.response.comment.CommentOfferingInfoResponse
-import com.zzang.chongdae.data.remote.dto.response.comment.CommentsResponse
-import com.zzang.chongdae.data.remote.dto.response.comment.UpdatedStatusResponse
-import com.zzang.chongdae.data.remote.dto.response.commentroom.CommentRoomsResponse
+import com.zzang.chongdae.remote.dto.request.CommentRequest
+import com.zzang.chongdae.remote.dto.response.comment.CommentOfferingInfoResponse
+import com.zzang.chongdae.remote.dto.response.comment.CommentsResponse
+import com.zzang.chongdae.remote.dto.response.comment.UpdatedStatusResponse
+import com.zzang.chongdae.remote.dto.response.commentroom.CommentRoomsResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
