+++ android/app/src/main/java/com/zzang/chongdae/remote/api/OfferingApiService.kt
@@ -1,13 +1,13 @@
 package com.zzang.chongdae.remote.api
 
-import com.zzang.chongdae.data.remote.dto.request.OfferingWriteRequest
-import com.zzang.chongdae.data.remote.dto.request.ProductUrlRequest
-import com.zzang.chongdae.data.remote.dto.response.offering.FiltersResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.MeetingsResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.OfferingDetailResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.OfferingsResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.ProductUrlResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.RemoteOffering
+import com.zzang.chongdae.remote.dto.request.OfferingWriteRequest
+import com.zzang.chongdae.remote.dto.request.ProductUrlRequest
+import com.zzang.chongdae.remote.dto.response.offering.FiltersResponse
+import com.zzang.chongdae.remote.dto.response.offering.MeetingsResponse
+import com.zzang.chongdae.remote.dto.response.offering.OfferingDetailResponse
+import com.zzang.chongdae.remote.dto.response.offering.OfferingsResponse
+import com.zzang.chongdae.remote.dto.response.offering.ProductUrlResponse
+import com.zzang.chongdae.remote.dto.response.offering.RemoteOffering
 import okhttp3.MultipartBody
 import retrofit2.Response
 import retrofit2.http.Body
