+++ android/app/src/main/java/com/zzang/chongdae/remote/mapper/FilterMapper.kt
@@ -2,12 +2,12 @@
 
 package com.zzang.chongdae.remote.mapper
 
-import com.zzang.chongdae.data.remote.dto.response.offering.RemoteFilter
-import com.zzang.chongdae.data.remote.dto.response.offering.RemoteFilterName
-import com.zzang.chongdae.data.remote.dto.response.offering.RemoteFilterType
 import com.zzang.chongdae.domain.model.Filter
 import com.zzang.chongdae.domain.model.FilterName
 import com.zzang.chongdae.domain.model.FilterType
+import com.zzang.chongdae.remote.dto.response.offering.RemoteFilter
+import com.zzang.chongdae.remote.dto.response.offering.RemoteFilterName
+import com.zzang.chongdae.remote.dto.response.offering.RemoteFilterType
 
 fun RemoteFilter.toDomain() =
     Filter(
