+++ android/app/src/main/java/com/zzang/chongdae/remote/mapper/MeetingsResponseMapper.kt
@@ -1,7 +1,7 @@
 package com.zzang.chongdae.remote.mapper
 
-import com.zzang.chongdae.data.remote.dto.response.offering.MeetingsResponse
 import com.zzang.chongdae.domain.model.Meetings
+import com.zzang.chongdae.remote.dto.response.offering.MeetingsResponse
 
 fun MeetingsResponse.toDomain() =
     Meetings(
