+++ android/app/src/main/java/com/zzang/chongdae/remote/source/AuthRemoteDataSourceImpl.kt
@@ -1,12 +1,12 @@
 package com.zzang.chongdae.remote.source
 
-import com.zzang.chongdae.remote.api.AuthApiService
-import com.zzang.chongdae.data.remote.dto.request.AccessTokenRequest
-import com.zzang.chongdae.data.remote.dto.response.auth.MemberResponse
-import com.zzang.chongdae.data.remote.util.safeApiCall
 import com.zzang.chongdae.data.source.AuthRemoteDataSource
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.api.AuthApiService
+import com.zzang.chongdae.remote.dto.request.AccessTokenRequest
+import com.zzang.chongdae.remote.dto.response.auth.MemberResponse
+import com.zzang.chongdae.remote.util.safeApiCall
 
 class AuthRemoteDataSourceImpl(
     private val service: com.zzang.chongdae.remote.api.AuthApiService,
