+++ android/app/src/main/java/com/zzang/chongdae/remote/source/CommentRemoteDataSourceImpl.kt
@@ -1,14 +1,13 @@
 package com.zzang.chongdae.remote.source
 
-import com.zzang.chongdae.remote.api.CommentApiService
-import com.zzang.chongdae.data.remote.dto.request.CommentRequest
-import com.zzang.chongdae.data.remote.dto.response.comment.CommentOfferingInfoResponse
-import com.zzang.chongdae.data.remote.dto.response.comment.CommentsResponse
-import com.zzang.chongdae.data.remote.dto.response.comment.UpdatedStatusResponse
-import com.zzang.chongdae.data.remote.util.safeApiCall
 import com.zzang.chongdae.data.source.comment.CommentRemoteDataSource
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.dto.request.CommentRequest
+import com.zzang.chongdae.remote.dto.response.comment.CommentOfferingInfoResponse
+import com.zzang.chongdae.remote.dto.response.comment.CommentsResponse
+import com.zzang.chongdae.remote.dto.response.comment.UpdatedStatusResponse
+import com.zzang.chongdae.remote.util.safeApiCall
 
 class CommentRemoteDataSourceImpl(
     private val service: com.zzang.chongdae.remote.api.CommentApiService,
