+++ android/app/src/main/java/com/zzang/chongdae/remote/source/OfferingDetailDataSourceImpl.kt
@@ -1,13 +1,13 @@
 package com.zzang.chongdae.remote.source
 
-import com.zzang.chongdae.data.remote.api.OfferingApiService
-import com.zzang.chongdae.data.remote.api.ParticipationApiService
-import com.zzang.chongdae.data.remote.dto.request.ParticipationRequest
-import com.zzang.chongdae.data.remote.dto.response.offering.OfferingDetailResponse
-import com.zzang.chongdae.data.remote.util.safeApiCall
 import com.zzang.chongdae.data.source.OfferingDetailDataSource
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.api.OfferingApiService
+import com.zzang.chongdae.remote.api.ParticipationApiService
+import com.zzang.chongdae.remote.dto.request.ParticipationRequest
+import com.zzang.chongdae.remote.dto.response.offering.OfferingDetailResponse
+import com.zzang.chongdae.remote.util.safeApiCall
 
 class OfferingDetailDataSourceImpl(
     private val offeringApiService: OfferingApiService,
