+++ android/app/src/main/java/com/zzang/chongdae/remote/source/OfferingRemoteDataSourceImpl.kt
@@ -1,17 +1,17 @@
 package com.zzang.chongdae.remote.source
 
-import com.zzang.chongdae.data.remote.api.OfferingApiService
-import com.zzang.chongdae.data.remote.dto.request.OfferingWriteRequest
-import com.zzang.chongdae.data.remote.dto.response.offering.FiltersResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.MeetingsResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.OfferingsResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.ProductUrlResponse
-import com.zzang.chongdae.data.remote.dto.response.offering.RemoteOffering
-import com.zzang.chongdae.data.remote.mapper.toProductUrlRequest
-import com.zzang.chongdae.data.remote.util.safeApiCall
 import com.zzang.chongdae.data.source.offering.OfferingRemoteDataSource
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.api.OfferingApiService
+import com.zzang.chongdae.remote.dto.request.OfferingWriteRequest
+import com.zzang.chongdae.remote.dto.response.offering.FiltersResponse
+import com.zzang.chongdae.remote.dto.response.offering.MeetingsResponse
+import com.zzang.chongdae.remote.dto.response.offering.OfferingsResponse
+import com.zzang.chongdae.remote.dto.response.offering.ProductUrlResponse
+import com.zzang.chongdae.remote.dto.response.offering.RemoteOffering
+import com.zzang.chongdae.remote.mapper.toProductUrlRequest
+import com.zzang.chongdae.remote.util.safeApiCall
 import okhttp3.MultipartBody
 
 class OfferingRemoteDataSourceImpl(
