+++ android/app/src/main/java/com/zzang/chongdae/remote/source/ParticipantRemoteDataSourceImpl.kt
@@ -1,11 +1,11 @@
 package com.zzang.chongdae.remote.source
 
-import com.zzang.chongdae.data.remote.api.ParticipationApiService
-import com.zzang.chongdae.data.remote.dto.response.participants.ParticipantsResponse
-import com.zzang.chongdae.data.remote.util.safeApiCall
 import com.zzang.chongdae.data.source.ParticipantRemoteDataSource
 import com.zzang.chongdae.domain.util.DataError
 import com.zzang.chongdae.domain.util.Result
+import com.zzang.chongdae.remote.api.ParticipationApiService
+import com.zzang.chongdae.remote.dto.response.participants.ParticipantsResponse
+import com.zzang.chongdae.remote.util.safeApiCall
 
 class ParticipantRemoteDataSourceImpl(
     private val service: ParticipationApiService,
